test(infinite-scroll-area): recategorize and clean up browser tests
